requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
pydantic-settings==2.8.1

httpx==0.28.1
python-dotenv==1.0.1

azure-search-documents==11.6.0
openai==1.68.2
azure-monitor-opentelemetry==1.6.5

pytest==8.3.5